Moves extracting hero indexes to a FeatureExtractor class.
This does a few things. It sets up a director "feature_extractors" where we can build a variety of feature extractor classes. The first one just emits the index of a hero in the list of heros. We'll want to upgrade this to emit other attributes of heros. We will probably want a "Hero" class we can use to do this, as we'll probably want to use it in other places.

This also moves PrototypeML to use the feature extractor, which simplifies the code a bit.