Adds oracle features.
If they're enabled, the models are able to take advantage. I'm not sure how the output of the Decision Tree works - I don't think it's right.

The Logistic Regression score falls  from .98 with only the oracle to like .5 with all the other features, but the decision tree seems to be correctly only using the oracle features. More to debug here.